revise sxrd
